第19次提交、完善分组、设备参数

+++ serverv2/src/main/java/com/kiwihouse/dto/ImprovedSetParamDto.java
@@ -17,7 +17,7 @@ import lombok.experimental.Accessors;
 public class ImprovedSetParamDto {
 
     private Double CT;
-    private Double Beep;
+    private Integer Beep;
     private Double volH;
     private Double leakH;
     private Double curH;
